feat: added rideDetails to EndRide API
- Added rideDetails to EndRide API
- Changed whenWithLocationUpdatesLock to return data with a -> a
