+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Action/UI/Ride.hs
@@ -13,8 +13,7 @@
 -}
 
 module Domain.Action.UI.Ride
-  ( DriverRideRes (..),
-    DriverRideListRes (..),
+  ( DriverRideListRes (..),
     OTPRideReq (..),
     UploadOdometerReq (..),
     UploadOdometerResp (..),
@@ -39,26 +38,14 @@ import Data.Ord
 import qualified Data.Text as T hiding (count, map)
 import qualified Data.Text as Text
 import Data.Time (Day)
-import Domain.Action.Dashboard.Ride
-import qualified Domain.Action.UI.Location as DLoc
+import qualified Domain.Action.UI.Ride.Common as RideCommon
 import qualified Domain.Action.UI.RideDetails as RD
-import qualified Domain.Types as DTC
-import qualified Domain.Types as DVST
-import qualified Domain.Types.BapMetadata as DSM
 import qualified Domain.Types.Booking as DRB
-import qualified Domain.Types.BookingCancellationReason as DBCR
-import qualified Domain.Types.DeliveryDetails as DParcel
-import qualified Domain.Types.DriverGoHomeRequest as DDGR
-import qualified Domain.Types.DriverInformation as DI
-import qualified Domain.Types.Exophone as DExophone
 import qualified Domain.Types.Location as DLoc
 import qualified Domain.Types.MerchantOperatingCity as DMOC
 import qualified Domain.Types.Person as DP
-import qualified Domain.Types.Rating as DRating
 import qualified Domain.Types.Ride as DRide
-import qualified Domain.Types.RideDetails as RD
 import qualified Domain.Types.StopInformation as DSI
-import qualified Domain.Types.VehicleVariant as DVeh
 import Environment
 import qualified EulerHS.Language as L
 import EulerHS.Prelude (withFile)
@@ -79,7 +66,6 @@ import Kernel.Streaming.Kafka.Producer.Types (KafkaProducerTools)
 import Kernel.Types.APISuccess
 import qualified Kernel.Types.Beckn.Domain as Domain
 import Kernel.Types.Common
-import Kernel.Types.Confidence
 import Kernel.Types.Id
 import Kernel.Utils.CalculateDistance (distanceBetweenInMeters)
 import Kernel.Utils.Common
@@ -87,7 +73,6 @@ import Kernel.Utils.Error.BaseError.HTTPError.BecknAPIError
 import qualified SharedLogic.Booking as SBooking
 import qualified SharedLogic.CallBAP as BP
 import qualified SharedLogic.CallBAPInternal as CallBAPInternal
-import SharedLogic.FareCalculator (fareSum)
 import SharedLogic.Ride
 import Storage.Beam.IssueManagement ()
 import Storage.Cac.TransporterConfig as SCTC
@@ -98,7 +83,6 @@ import qualified Storage.CachedQueries.Merchant.Overlay as CMP
 import qualified Storage.CachedQueries.ValueAddNP as CQVAN
 import qualified Storage.CachedQueries.VehicleServiceTier as CQVST
 import qualified Storage.Queries.Booking as QBooking
-import qualified Storage.Queries.BookingCancellationReason as QBCR
 import qualified Storage.Queries.DriverInformation as QDI
 import qualified Storage.Queries.FleetDriverAssociation as FDV
 import qualified Storage.Queries.Location as QLoc
@@ -170,112 +154,6 @@ newtype UploadOdometerResp = UploadOdometerResp
   deriving stock (Eq, Show, Generic)
   deriving anyclass (ToJSON, FromJSON, ToSchema)
 
-data BookingType = CURRENT | ADVANCED
-  deriving stock (Eq, Show, Generic, Ord)
-  deriving anyclass (ToJSON, FromJSON, ToSchema)
-
-data DriverRideRes = DriverRideRes
-  { id :: Id DRide.Ride,
-    shortRideId :: ShortId DRide.Ride,
-    status :: DRide.RideStatus,
-    fromLocation :: DLoc.LocationAPIEntity,
-    toLocation :: Maybe DLoc.LocationAPIEntity,
-    stops :: [Stop],
-    driverName :: Text,
-    driverNumber :: Maybe Text,
-    vehicleVariant :: DVeh.VehicleVariant,
-    pickupDropOutsideOfThreshold :: Maybe Bool,
-    vehicleModel :: Text,
-    vehicleColor :: Text,
-    vehicleNumber :: Text,
-    computedFare :: Maybe Money,
-    estimatedBaseFare :: Money,
-    computedFareWithCurrency :: Maybe PriceAPIEntity,
-    estimatedBaseFareWithCurrency :: PriceAPIEntity,
-    estimatedDistance :: Maybe Meters,
-    estimatedDistanceWithUnit :: Maybe Distance,
-    driverSelectedFare :: Money,
-    driverSelectedFareWithCurrency :: PriceAPIEntity,
-    actualRideDistance :: HighPrecMeters,
-    actualRideDistanceWithUnit :: Distance,
-    rideRating :: Maybe Int,
-    riderName :: Maybe Text,
-    tripStartTime :: Maybe UTCTime,
-    tripEndTime :: Maybe UTCTime,
-    specialLocationTag :: Maybe Text,
-    actualDuration :: Maybe Seconds,
-    estimatedDuration :: Maybe Seconds,
-    chargeableDistance :: Maybe Meters,
-    chargeableDistanceWithUnit :: Maybe Distance,
-    exoPhone :: Text,
-    bapName :: Maybe Text,
-    bapLogo :: Maybe BaseUrl,
-    createdAt :: UTCTime,
-    updatedAt :: UTCTime,
-    customerExtraFee :: Maybe Money,
-    customerExtraFeeWithCurrency :: Maybe PriceAPIEntity,
-    disabilityTag :: Maybe Text,
-    coinsRewardedOnGoldTierRide :: Maybe Int,
-    requestedVehicleVariant :: DVeh.VehicleVariant,
-    isOdometerReadingsRequired :: Bool,
-    vehicleServiceTier :: DVST.ServiceTierType,
-    vehicleServiceTierName :: Text,
-    isVehicleAirConditioned :: Maybe Bool,
-    vehicleCapacity :: Maybe Int,
-    driverGoHomeRequestId :: Maybe (Id DDGR.DriverGoHomeRequest),
-    payerVpa :: Maybe Text,
-    autoPayStatus :: Maybe DI.DriverAutoPayStatus,
-    customerCancellationDues :: HighPrecMoney,
-    estimatedTollCharges :: Maybe HighPrecMoney,
-    parkingCharge :: Maybe HighPrecMoney,
-    tollCharges :: Maybe HighPrecMoney,
-    tollConfidence :: Maybe Confidence,
-    customerCancellationDuesWithCurrency :: PriceAPIEntity,
-    estimatedTollChargesWithCurrency :: Maybe PriceAPIEntity,
-    parkingChargeWithCurrency :: Maybe PriceAPIEntity,
-    tollChargesWithCurrency :: Maybe PriceAPIEntity,
-    isFreeRide :: Maybe Bool,
-    stopLocationId :: Maybe (Id DLoc.Location),
-    tripCategory :: DTC.TripCategory,
-    nextStopLocation :: Maybe DLoc.Location,
-    lastStopLocation :: Maybe DLoc.Location,
-    startOdometerReading :: Maybe DRide.OdometerReading,
-    endOdometerReading :: Maybe DRide.OdometerReading,
-    returnTime :: Maybe UTCTime,
-    vehicleAge :: Maybe Months,
-    roundTrip :: Bool,
-    tripScheduledAt :: UTCTime,
-    isValueAddNP :: Bool,
-    bookingType :: BookingType,
-    enableFrequentLocationUpdates :: Maybe Bool,
-    fleetOwnerId :: Maybe Text,
-    enableOtpLessRide :: Bool,
-    cancellationSource :: Maybe DBCR.CancellationSource,
-    tipAmount :: Maybe PriceAPIEntity,
-    penalityCharge :: Maybe PriceAPIEntity,
-    senderDetails :: Maybe DeliveryPersonDetailsAPIEntity,
-    receiverDetails :: Maybe DeliveryPersonDetailsAPIEntity,
-    extraFareMitigationFlag :: Maybe Bool,
-    parcelType :: Maybe DParcel.ParcelType,
-    parcelQuantity :: Maybe Int,
-    isPetRide :: Bool,
-    isInsured :: Maybe Bool,
-    insuredAmount :: Maybe Text
-  }
-  deriving (Generic, Show, FromJSON, ToJSON, ToSchema)
-
-data Stop = Stop
-  { location :: DLoc.LocationAPIEntity,
-    stopInfo :: Maybe DSI.StopInformation
-  }
-  deriving (Generic, Show, FromJSON, ToJSON, ToSchema)
-
-data DeliveryPersonDetailsAPIEntity = DeliveryPersonDetailsAPIEntity
-  { name :: Text,
-    primaryExophone :: Text
-  }
-  deriving (Generic, Show, FromJSON, ToJSON, ToSchema)
-
 data CoinsEarned = CoinsEarned
   { coins :: Int,
     eventType :: Text
@@ -290,7 +168,7 @@ data OTPRideReq = OTPRideReq
   deriving (Generic, Show, FromJSON, ToJSON, ToSchema)
 
 newtype DriverRideListRes = DriverRideListRes
-  { list :: [DriverRideRes]
+  { list :: [RideCommon.DriverRideRes]
   }
   deriving (Generic, Show, FromJSON, ToJSON, ToSchema)
 
@@ -321,7 +199,7 @@ listDriverRides driverId mbLimit mbOffset mbOnlyActive mbRideStatus mbDay mbFlee
     isValueAddNP <- CQVAN.isValueAddNP booking.bapId
     stopsInfo <- if (fromMaybe False ride.hasStops) then QSI.findAllByRideId ride.id else return []
     let goHomeReqId = ride.driverGoHomeRequestId
-    mkDriverRideRes rideDetail driverNumber rideRating mbExophone (ride, booking) bapMetadata goHomeReqId (Just driverInfo) isValueAddNP stopsInfo
+    RideCommon.mkDriverRideRes rideDetail driverNumber rideRating mbExophone (ride, booking) bapMetadata goHomeReqId (Just driverInfo) isValueAddNP stopsInfo
   filteredRides <- case mbFleetOwnerId of
     Just fleetOwnerId -> do
       isFleetDriver <- FDV.findByDriverIdAndFleetOwnerId driverId fleetOwnerId True
@@ -329,142 +207,6 @@ listDriverRides driverId mbLimit mbOffset mbOnlyActive mbRideStatus mbDay mbFlee
       return $ filter (\ride -> ride.fleetOwnerId == (Just fleetOwnerId)) driverRideLis
     Nothing -> return driverRideLis
   pure . DriverRideListRes $ sortOn (Down . (.bookingType)) filteredRides
-
-mkDriverRideRes ::
-  ( EncFlow m r,
-    CacheFlow m r,
-    EsqDBFlow m r
-  ) =>
-  RD.RideDetails ->
-  Maybe Text ->
-  Maybe DRating.Rating ->
-  Maybe DExophone.Exophone ->
-  (DRide.Ride, DRB.Booking) ->
-  Maybe DSM.BapMetadata ->
-  Maybe (Id DDGR.DriverGoHomeRequest) ->
-  Maybe DI.DriverInformation ->
-  Bool ->
-  [DSI.StopInformation] ->
-  m DriverRideRes
-mkDriverRideRes rideDetails driverNumber rideRating mbExophone (ride, booking) bapMetadata goHomeReqId driverInfo isValueAddNP stopsInfo = do
-  let fareParams = booking.fareParams
-      estimatedBaseFare = fareSum (fareParams{driverSelectedFare = Nothing}) Nothing -- it should not be part of estimatedBaseFare
-  let initial = "" :: Text
-  (nextStopLocation, lastStopLocation) <- case booking.tripCategory of
-    DTC.Rental _ -> calculateLocations booking.id booking.stopLocationId
-    _ -> return (Nothing, Nothing)
-  cancellationReason <- if ride.status == DRide.CANCELLED then runInReplica (QBCR.findByRideId (Just ride.id)) else pure Nothing
-  return $
-    DriverRideRes
-      { id = ride.id,
-        shortRideId = ride.shortId,
-        status = ride.status,
-        fromLocation = DLoc.makeLocationAPIEntity booking.fromLocation,
-        toLocation = DLoc.makeLocationAPIEntity <$> booking.toLocation,
-        stops = map (makeStop stopsInfo) booking.stops,
-        driverName = rideDetails.driverName,
-        driverNumber,
-        vehicleNumber = rideDetails.vehicleNumber,
-        vehicleColor = fromMaybe initial rideDetails.vehicleColor,
-        vehicleVariant = fromMaybe DVeh.SEDAN rideDetails.vehicleVariant,
-        vehicleModel = fromMaybe initial rideDetails.vehicleModel,
-        computedFare = roundToIntegral <$> ride.fare,
-        computedFareWithCurrency = flip PriceAPIEntity ride.currency <$> ride.fare,
-        estimatedDuration = booking.estimatedDuration,
-        actualDuration = roundToIntegral <$> (diffUTCTime <$> ride.tripEndTime <*> ride.tripStartTime),
-        estimatedBaseFare = roundToIntegral estimatedBaseFare,
-        estimatedBaseFareWithCurrency = PriceAPIEntity estimatedBaseFare ride.currency,
-        estimatedDistance = booking.estimatedDistance,
-        estimatedDistanceWithUnit = convertMetersToDistance booking.distanceUnit <$> booking.estimatedDistance,
-        driverSelectedFare = roundToIntegral $ fromMaybe 0.0 fareParams.driverSelectedFare,
-        driverSelectedFareWithCurrency = flip PriceAPIEntity fareParams.currency $ fromMaybe 0.0 fareParams.driverSelectedFare,
-        actualRideDistance = ride.traveledDistance,
-        actualRideDistanceWithUnit = convertHighPrecMetersToDistance ride.distanceUnit ride.traveledDistance,
-        createdAt = ride.createdAt,
-        updatedAt = ride.updatedAt,
-        riderName = booking.riderName,
-        pickupDropOutsideOfThreshold = ride.pickupDropOutsideOfThreshold,
-        tripStartTime = ride.tripStartTime,
-        tripEndTime = ride.tripEndTime,
-        specialLocationTag = booking.specialLocationTag,
-        rideRating = rideRating <&> (.ratingValue),
-        chargeableDistance = ride.chargeableDistance,
-        chargeableDistanceWithUnit = convertMetersToDistance ride.distanceUnit <$> ride.chargeableDistance,
-        exoPhone = maybe booking.primaryExophone (\exophone -> if not exophone.isPrimaryDown then exophone.primaryPhone else exophone.backupPhone) mbExophone,
-        customerExtraFee = roundToIntegral <$> fareParams.customerExtraFee,
-        customerExtraFeeWithCurrency = flip PriceAPIEntity fareParams.currency <$> fareParams.customerExtraFee,
-        bapName = bapMetadata <&> (.name),
-        bapLogo = bapMetadata >>= (.logoUrl),
-        disabilityTag = booking.disabilityTag,
-        coinsRewardedOnGoldTierRide = booking.coinsRewardedOnGoldTierRide,
-        requestedVehicleVariant = DVeh.castServiceTierToVariant booking.vehicleServiceTier,
-        isOdometerReadingsRequired = DTC.isOdometerReadingsRequired booking.tripCategory,
-        vehicleServiceTier = booking.vehicleServiceTier,
-        vehicleServiceTierName = booking.vehicleServiceTierName,
-        vehicleCapacity = booking.vehicleServiceTierSeatingCapacity,
-        isVehicleAirConditioned = booking.isAirConditioned,
-        driverGoHomeRequestId = goHomeReqId,
-        payerVpa = driverInfo >>= (.payerVpa),
-        autoPayStatus = driverInfo >>= (.autoPayStatus),
-        isFreeRide = ride.isFreeRide,
-        customerCancellationDues = fromMaybe 0 fareParams.customerCancellationDues,
-        estimatedTollCharges = fareParams.tollCharges,
-        parkingCharge = fareParams.parkingCharge,
-        tollCharges = ride.tollCharges,
-        tollConfidence = ride.tollConfidence,
-        customerCancellationDuesWithCurrency = flip PriceAPIEntity fareParams.currency $ fromMaybe 0 fareParams.customerCancellationDues,
-        estimatedTollChargesWithCurrency = flip PriceAPIEntity fareParams.currency <$> fareParams.tollCharges,
-        parkingChargeWithCurrency = flip PriceAPIEntity fareParams.currency <$> fareParams.parkingCharge,
-        tollChargesWithCurrency = flip PriceAPIEntity ride.currency <$> ride.tollCharges,
-        startOdometerReading = ride.startOdometerReading,
-        endOdometerReading = ride.endOdometerReading,
-        stopLocationId = booking.stopLocationId,
-        tripCategory = booking.tripCategory,
-        returnTime = booking.returnTime,
-        vehicleAge = rideDetails.vehicleAge,
-        roundTrip = fromMaybe False booking.roundTrip,
-        nextStopLocation = nextStopLocation,
-        lastStopLocation = lastStopLocation,
-        tripScheduledAt = booking.startTime,
-        bookingType = if ride.status == DRide.NEW && ride.isAdvanceBooking && maybe False (.hasAdvanceBooking) driverInfo then ADVANCED else CURRENT,
-        isValueAddNP,
-        enableFrequentLocationUpdates = ride.enableFrequentLocationUpdates,
-        fleetOwnerId = rideDetails.fleetOwnerId,
-        enableOtpLessRide = fromMaybe False ride.enableOtpLessRide,
-        cancellationSource = fmap (\cr -> cr.source) cancellationReason,
-        tipAmount = flip PriceAPIEntity ride.currency <$> ride.tipAmount,
-        penalityCharge = flip PriceAPIEntity ride.currency <$> ride.cancellationFeeIfCancelled,
-        senderDetails = booking.senderDetails <&> (\sd -> DeliveryPersonDetailsAPIEntity (sd.name) sd.primaryExophone),
-        receiverDetails = booking.receiverDetails <&> (\rd -> DeliveryPersonDetailsAPIEntity (rd.name) rd.primaryExophone),
-        extraFareMitigationFlag = driverInfo >>= (.extraFareMitigationFlag),
-        parcelType = booking.parcelType,
-        parcelQuantity = booking.parcelQuantity,
-        isPetRide = booking.isPetRide,
-        isInsured = Just $ ride.isInsured,
-        insuredAmount = ride.insuredAmount
-      }
-
-makeStop :: [DSI.StopInformation] -> DLoc.Location -> Stop
-makeStop stopsInfo loc = do
-  Stop (DLoc.makeLocationAPIEntity loc) (find (\stop -> stop.stopLocId == loc.id) stopsInfo)
-
-calculateLocations ::
-  ( CacheFlow m r,
-    EsqDBFlow m r
-  ) =>
-  Id DRB.Booking ->
-  Maybe (Id DLoc.Location) ->
-  m (Maybe DLoc.Location, Maybe DLoc.Location)
-calculateLocations bookingId stopLocationId = do
-  maxOrder <- QLM.maxOrderByEntity bookingId.getId
-  case stopLocationId of
-    Nothing -> do
-      lastLoc <- if maxOrder == 0 then pure Nothing else mkLocationFromLocationMapping bookingId.getId maxOrder
-      return (Nothing, lastLoc)
-    Just nextStopId -> do
-      nextLoc <- QLoc.findById nextStopId
-      lastLoc <- mkLocationFromLocationMapping bookingId.getId (maxOrder - 1)
-      return (nextLoc, lastLoc)
 
 arrivedAtPickup :: (EncFlow m r, CacheFlow m r, EsqDBFlow m r, EsqDBReplicaFlow m r, HasShortDurationRetryCfg r c, HasFlowEnv m r '["nwAddress" ::: BaseUrl], HasHttpClientOptions r c, HasFlowEnv m r '["internalEndPointHashMap" ::: HM.HashMap BaseUrl BaseUrl], HasFlowEnv m r '["kafkaProducerTools" ::: KafkaProducerTools], HasFlowEnv m r '["ondcTokenHashMap" ::: HM.HashMap KeyConfig TokenConfig]) => Id DRide.Ride -> LatLong -> m APISuccess
 arrivedAtPickup rideId req = do
@@ -495,7 +237,7 @@ arrivedAtPickup rideId req = do
       TN.sendOverlay moCityId driver $ TN.mkOverlayReq overlay
       QDI.updateExtraFareMitigation (Just False) driverId
 
-otpRideCreate :: DP.Person -> Text -> DRB.Booking -> Maybe Text -> Flow DriverRideRes
+otpRideCreate :: DP.Person -> Text -> DRB.Booking -> Maybe Text -> Flow RideCommon.DriverRideRes
 otpRideCreate driver otpCode booking clientId = do
   transporter <-
     QM.findById booking.providerId
@@ -519,7 +261,7 @@ otpRideCreate driver otpCode booking clientId = do
   mbExophone <- CQExophone.findByPrimaryPhone booking.primaryExophone
   bapMetadata <- CQSM.findBySubscriberIdAndDomain (Id booking.bapId) Domain.MOBILITY
   isValueAddNP <- CQVAN.isValueAddNP booking.bapId
-  mkDriverRideRes rideDetails driverNumber Nothing mbExophone (ride, booking) bapMetadata ride.driverGoHomeRequestId Nothing isValueAddNP stopsInfo
+  RideCommon.mkDriverRideRes rideDetails driverNumber Nothing mbExophone (ride, booking) bapMetadata ride.driverGoHomeRequestId Nothing isValueAddNP stopsInfo
   where
     errHandler uBooking transporter exc
       | Just BecknAPICallError {} <- fromException @BecknAPICallError exc = SBooking.cancelBooking uBooking (Just driver) transporter >> throwM exc
